% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bess.cun.r
\name{bess.cun}
\alias{bess.cun}
\title{Cylindrical Bessel function of order one \eqn{J_1(x)}.}
\usage{
bess.cun(x)
}
\description{
Cylindrical Bessel function of order one \eqn{J_1(x)}.
}
\details{
Cylindrical Bessel function calculated by series summation.
}

