% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bess.sun.r
\name{bess.sun}
\alias{bess.sun}
\title{Spherical Bessel function of order one \eqn{j_1(x)}}
\usage{
bess.sun(x)
}
\arguments{
\item{x}{Complex argument of the Spherical Bessel function of order one.}
}
\value{
Complex values for \eqn{j_1(x)}.
}
\description{
Spherical Bessel function of order one \eqn{j_1(x)}
}
\details{
It uses the series solution to calculate the values of
the Spherical Bessel function for \eqn{x \to 0}.
}
\examples{
print(bess.sun(3+2i))
print(bess.sun(0:10))
}

