% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bess.szr.r
\name{bess.szr}
\alias{bess.szr}
\title{Spherical Bessel function of order one \eqn{j_0(x)}}
\usage{
bess.szr(x)
}
\arguments{
\item{x}{Complex argument of the Spherical Bessel function of order one.}
}
\value{
Complex values for \eqn{j_0(x)}.
}
\description{
Spherical Bessel function of order one \eqn{j_0(x)}
}
\details{
It uses the series solution to calculate the values of
the Spherical Bessel function for \eqn{x \to 0}.
}
\examples{
print(bess.szr(3+2i))
print(bess.szr(0:10))
}

