% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comp.neu.r
\name{comp.neu}
\alias{comp.neu}
\title{Compare results for Cylindrical Neumann Functions.}
\usage{
comp.neu(n, x)
}
\arguments{
\item{n}{The order of the Ricatti-Bessel function.}

\item{x}{The argument of \eqn{\psi_n(x)=xj_n(x)}.}
}
\description{
Compare results for Cylindrical Neumann Functions.
}
\details{
Compare results using vswf, built in \code{R} and \code{gsl} algorithms.
}
\examples{
x<-5
nmax<-10
print(comp.neu(5,3))
}

