\name{lcfe.afs}
\alias{lcfe.afs}
\title{Auxiliary function \eqn{S_n(x)=n/x}. Used on calculations of Bessel functions.}
\usage{
lcfe.afs(n, x)
}
\arguments{
  \item{n}{The order of the Bessel function.}

  \item{x}{The argument of the Bessel function.}
}
\value{
The value \code{n/x}.
}
\description{
Auxiliary function \eqn{S_n(x)=n/x}. Used on calculations
of Bessel functions.
}

