% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmie.ofc.r
\name{lmie.ofc}
\alias{lmie.ofc}
\title{Lorentz-Mie Optical Force Calculation.}
\usage{
lmie.ofc(m, x, GTE, GTM, lmax = floor(x + 2 + 4 * x^(1/3)), code = "C")
}
\arguments{
\item{m}{The ratio between the refractive indices.}

\item{x}{The form factor value.}

\item{by}{\code{LD} using logarithmic derivative; \code{RB} using the ratio
between Ricatti-Bessel functions.}
}
\value{
The coefficients \eqn{A_n}, \eqn{B_n} and \eqn{C_n} used to calculate
the Optical force.
}
\description{
Lorentz-Mie Optical Force Calculation.
}
\examples{
lmax<-20
lambda=.532  # Green 532 nm
m<-ridx.pol(lambda)/ridx.h2o(lambda)
g<-vswf.mpw(lmax+1)
g<-vswf.bbp(3,4,1,2,3,lmax+1,1,1,1)
x<-20

F<-lmie.ofc(m,x,g$GTE,g$GTM,lmax)
print(F)
}

