% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmie.rfc.r
\name{lmie.rfc}
\alias{lmie.rfc}
\title{n<-1.1;thmax<-79
n<-1.2;thmax<-72
#n<-1.4;thmax<-64
#n<-1.6;thmax<-60
#n<-1.8;thmax<-59
#n<-2.0;thmax<-59
#n<-2.5;thmax<-64}
\usage{
lmie.rfc(th.i, n, pp = 0.5)
}
\description{
th<-seq(0,90,.1)*pi/180
a<-lmie.rfc(th,1/n,pp=0)
b<-lmie.rfc(th,1/n,pp=1)
}
\details{
tg<-th*180/pi
plot(tg,a$Ft,type='l',xlab=expression(theta),ylab="Q",ylim=range(c(a$Ft,b$Ft)))
points(tg,abs(a$Fg),type='l',col='blue')
points(tg,abs(a$Fs),type='l',col='red')

points(tg,abs(b$Ft),lty=2,type='l')
points(tg,abs(b$Fg),lty=2,type='l',col='blue')
points(tg,abs(b$Fs),lty=2,type='l',col='red')

legend("topleft",c("Full S","Grad S","Scat S","Full P","Grad P","Scat P"),
      lty=c(1,1,1,2,2,2),col=c("black","blue","red"))

#abline(v=thmax,col='green')
#print(round(min(a$Fg),3))
}

