\name{reff.cdh}
\alias{reff.cdh}
\title{Reference function Derivative of Cylindrical Hankel function (1,2).}
\usage{
reff.cdh(n, x, type = 1)
}
\arguments{
  \item{x}{The argument of the function}

  \item{n}{The order of the function}

  \item{type}{1 or 2.}
}
\description{
Reference function Derivative of Cylindrical Hankel
function (1,2).
}
\details{
The derivative of the Cylindrical Hankel function given by
\eqn{h_n^{(1,2)}=j_n(x)\pm y_n(x)}.
}

