% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reff.cdj.r
\name{reff.cdj}
\alias{reff.cdj}
\title{Refference function Derivative of the Spherical Bessel function.}
\usage{
reff.cdj(n, x)
}
\arguments{
\item{n}{The order of the function}

\item{x}{The argument of the function}
}
\value{
The value of the derivative
}
\description{
Refference function Derivative of the Spherical Bessel function.
}
\details{
Uses the definition and built-in \code{R} function to give the
}

