% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reff.chn.r
\name{reff.chn}
\alias{reff.chn}
\title{Reference function Cylindrical Hankel function (1,2).}
\usage{
reff.chn(x, n, type = 1)
}
\arguments{
\item{x}{The argument of the function}

\item{n}{The order of the function}

\item{type}{1 or 2.}
}
\description{
Reference function Cylindrical Hankel function (1,2).
}
\details{
The Cylindrical Hankel
function given by \eqn{h_n^{(1,2)}=j_n(x)\pm y_n(x)}.
}

