% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reff.rdj.r
\name{reff.rdj}
\alias{reff.rdj}
\title{Reference function Derivative of Ricatti-Bessel function .}
\usage{
reff.rdj(n, x)
}
\arguments{
\item{n}{The order of the function}

\item{x}{The argument of the function}
}
\value{
The derivative \eqn{\psi_n'(x)=j_n(x)+xj_n'(x)}.
}
\description{
Reference function Derivative of Ricatti-Bessel function .
}
\details{
The derivative of the Ricatti-Bessel function \eqn{\psi_n(x)=xj_n(x)}.
}

