% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reff.rjn.r
\name{reff.rjn}
\alias{reff.rjn}
\title{Ricatti-Bessel function}
\usage{
reff.rjn(n, x)
}
\arguments{
\item{n}{The order of the function}

\item{x}{The argument of the funcion}
}
\value{
The Ricatti-Bessel function \eqn{\psi_n(x)=xj_n(x)}
}
\description{
Ricatti-Bessel function
}
\details{
Only a parser to the built-in \code{R} function.
}

