% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reff.sdj.r
\name{reff.sdj}
\alias{reff.sdj}
\title{Reference function Derivative of Spherical Bessel function .}
\usage{
reff.sdj(x, n)
}
\arguments{
\item{x}{The argument of the function}

\item{n}{The order of the function}
}
\value{
The derivative \eqn{j_n'(x)}.
}
\description{
Reference function Derivative of Spherical Bessel function .
}
\details{
The derivative of the Spherical Bessel function \eqn{j_n(x)}.
}

