% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridx.pol.r
\name{ridx.pol}
\alias{ridx.pol}
\title{Parameters: Refraction index of poliestirene.}
\usage{
ridx.pol(l)
}
\arguments{
\item{l}{The wavelength in micrometers}
}
\value{
The index of refraction of the water as function of the wavelength
}
\description{
Parameters: Refraction index of poliestirene.
}
\examples{
l<-seq(.38,.75,.01)
u<-ridx.pol(l)
plot(l,u,type='l')
}
\seealso{
\code{\link{ridx.h2o}}
}

