% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.abc.r
\name{test.abc}
\alias{test.abc}
\title{Test the calculation}
\usage{
test.abc(n, x)
}
\arguments{
\item{n}{The order of the function}

\item{x}{The argument of the function}
}
\value{
Data frame with values \eqn{A_n}, \eqn{B_n} and \eqn{C_n}.
}
\description{
Test the calculation
}
\details{
Must be rewritten
}

