% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comp.cbd.r
\name{comp.cbd}
\alias{comp.cbd}
\title{Checks the results of Lentz method and internal R Bessel function}
\usage{
comp.cbd(n, x, code = "C")
}
\arguments{
\item{n}{Order of the ratio given by \eqn{\gamma_n=J_n/J_{n+1}}{g[n]=J_n/J_{n+1}}.}

\item{x}{Argument of Bessel functions.}

\item{code}{If C or native R function.}
}
\value{
Data frame with the values calculated by the algorithms.
}
\description{
Checks the results of Lentz method and internal R Bessel function
}
\details{
In order to check the results of Lentz method for calculation
of Cylindrical Bessel ratio of consecutive function \eqn{\gamma_n}{g[n]}.
}
\examples{
comp.cbd(5,4,code="C")
comp.cbd(5,4,code="R")
}
\seealso{
\code{\link{lcfe.cbi}}, \code{\link{lcfe.cbl}}, \code{\link{lcfe.cbd}},
\code{\link{lcfe.afs}}, \code{\link{besselJ}}.
}

