% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lcfe.afs.r
\name{lcfe.afs}
\alias{lcfe.afs}
\title{Auxiliary function \eqn{S_n(x)=n/x}. Used on calculations of Bessel functions.}
\usage{
lcfe.afs(n, x)
}
\arguments{
\item{n}{The order of the Bessel function.}

\item{x}{The argument of the Bessel function.}
}
\value{
The value \code{n/x}.
}
\description{
Auxiliary function \eqn{S_n(x)=n/x}. Used on calculations of Bessel functions.
}
\examples{
u<-0:100
n<-sample(u,10)
x<-sample(u,10)
print(data.frame(n,x,"n/x"=n/x,"lcfe.afs"=lcfe.afs(n,x)))
}

