\name{lmie.ofc}
\alias{lmie.ofc}
\title{Lorentz-Mie Optical Force Calculation.}
\usage{
lmie.ofc(m, x)
}
\arguments{
  \item{m}{The ratio between the refractive indices.}

  \item{x}{The form factor value.}

  \item{by}{\code{LD} using logarithmic derivative;
  \code{RB} using the ratio between Ricatti-Bessel
  functions.}
}
\value{
The coefficients \eqn{A_n}, \eqn{B_n} and \eqn{C_n} used to
calculate the Optical force.
}
\description{
Lorentz-Mie Optical Force Calculation.
}

