\name{lmie.sct}
\alias{lmie.sct}
\title{Scattering coeficients in the Lorentz-Mie}
\usage{
lmie.sct(m, x)
}
\arguments{
  \item{m}{Ratio of refraction indices.}

  \item{x}{Form factor}
}
\value{
Extinction and Scattering coefficients
}
\description{
Scattering coeficients in the Lorentz-Mie
}
\details{
Scattering and Extinction coefficients.
}
\examples{
# Cantrell papers.
}

