\name{reff.cdj}
\alias{reff.cdj}
\title{Refference function Derivative of the Spherical Bessel function.}
\usage{
reff.cdj(n, x)
}
\arguments{
  \item{x}{The argument of the function}

  \item{n}{The order of the function}
}
\value{
The value of the derivative
}
\description{
Refference function Derivative of the Spherical Bessel
function.
}
\details{
Uses the definition and built-in \code{R} function to give
the
}

