% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reff.cjn.r
\name{reff.cjn}
\alias{reff.cjn}
\title{Cylindrical Bessel function}
\usage{
reff.cjn(n, x)
}
\arguments{
\item{n}{The order of the function}

\item{x}{The argument of the funcion}
}
\value{
The Cylindrical Bessel function
}
\description{
Cylindrical Bessel function
}
\details{
Only a parser to the built-in \code{R} function.
}

