\name{reff.sjn}
\alias{reff.sjn}
\title{Reference function for Spherical Bessel Functions.}
\usage{
reff.sjn(x, n)
}
\description{
Reference function for Spherical Bessel Functions.
}
\details{
It uses de built in Cylindrical Bessel function to
calculate the Spherical Bessel function, taking care about
the zero.
}

