\name{test.abn}
\alias{test.abn}
\title{Test the calculation of Mie Coefficients}
\usage{
test.abn(n, m, x)
}
\arguments{
  \item{n}{The order of the function}

  \item{m}{The relative index of refraction}

  \item{x}{The argument of the function}
}
\value{
Data frame with values \eqn{A_n}, \eqn{B_n} and \eqn{C_n}.
}
\description{
Test the calculation of Mie Coefficients
}

